vcpe_scenarios/vcpe/5gc_docker_events/docker_events_notifier/requirements.txt
flask
PyYAML
requests
docker-events